Layout is not looking good
Video part should be on left side and occupy 60% of screen size

chat window should be on right side and occupy 40% 0f screen size.

and use variable to give background and other colours.

Add 3 to 4 theme option and give a dropdown on the top of the window to choose theme